created shared functionalities with conditional symbols
