+++ Microsoft.Dynamics365.UIAutomation.Api.2016/Microsoft.Dynamics365.UIAutomation.Api.OnPremise.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{80A65998-62A8-4E7B-87A0-8492C16703D3}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Microsoft.Dynamics365.UIAutomation.Api._2016</RootNamespace>
-    <AssemblyName>Microsoft.Dynamics365.UIAutomation.Api.2016</AssemblyName>
+    <RootNamespace>Microsoft.Dynamics365.UIAutomation.Api._OnPremise</RootNamespace>
+    <AssemblyName>Microsoft.Dynamics365.UIAutomation.Api.OnPremise</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -17,7 +17,7 @@
     <DebugType>full</DebugType>
     <Optimize>false</Optimize>
     <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>TRACE;DEBUG;CRM_2016</DefineConstants>
+    <DefineConstants>TRACE;DEBUG;CRM_ONPREM</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
