+++ Microsoft.Dynamics365.UIAutomation.Api.365/packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net461" />
+  <package id="Selenium.Support" version="3.8.0" targetFramework="net461" />
+  <package id="Selenium.WebDriver" version="3.8.0" targetFramework="net461" />
+</packages>
