+++ Microsoft.Dynamics365.UIAutomation.Api.Shared/DTOs/XrmElementReference.cs
@@ -583,11 +583,19 @@ namespace Microsoft.Dynamics365.UIAutomation.Api
         }
         public static class GuidedHelp
         {
+#if CRM_ONPREM
+            public static string GuideIFrame = "InlineDialog_Iframe";
             public static string MarsOverlay = "InlineDialog_Background";
             public static string ButBegin = "GuidedHelp_ButBegin";
             public static string ButtonClose = "buttonClose";
             public static string Close = "GuidedHelp_Close";
-
+#endif
+#if CRM_ONLINE
+            public static string MarsOverlay = "GuidedHelp_MarsOverlay";
+            public static string ButBegin = "GuidedHelp_ButBegin";
+            public static string ButtonClose = "GuidedHelp_ButtonClose";
+            public static string Close = "GuidedHelp_Close";
+#endif
         }
         public static class Notification
         {
