+++ Microsoft.Dynamics365.UIAutomation.Browser/BrowserDriverFactory.cs
@@ -5,7 +5,6 @@ using OpenQA.Selenium;
 using OpenQA.Selenium.Chrome;
 using OpenQA.Selenium.Firefox;
 using OpenQA.Selenium.IE;
-using OpenQA.Selenium.PhantomJS;
 using OpenQA.Selenium.Support.Events;
 using OpenQA.Selenium.Edge;
 using System;
