+++ Microsoft.Dynamics365.UIAutomation.Browser/packages.config
@@ -2,7 +2,7 @@
 <!--Copyright (c) Microsoft Corporation. All rights reserved.
     Licensed under the MIT license.-->
 <packages>
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net46" />
-  <package id="Selenium.Support" version="3.5.1" targetFramework="net46" />
-  <package id="Selenium.WebDriver" version="3.5.1" targetFramework="net46" />
+  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net46" />
+  <package id="Selenium.Support" version="3.8.0" targetFramework="net46" />
+  <package id="Selenium.WebDriver" version="3.8.0" targetFramework="net46" />
 </packages>
