+++ Microsoft.Dynamics365.UIAutomation.Sample.2016/packages.config
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.ApplicationInsights" version="2.4.0" targetFramework="net461" />
+  <package id="MSTest.TestAdapter" version="1.2.0" targetFramework="net461" />
+  <package id="MSTest.TestFramework" version="1.2.0" targetFramework="net461" />
+  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net461" />
+  <package id="Selenium.Chrome.WebDriver" version="2.35" targetFramework="net461" />
+  <package id="Selenium.Support" version="3.8.0" targetFramework="net461" />
+  <package id="Selenium.WebDriver" version="3.8.0" targetFramework="net461" />
+  <package id="System.Diagnostics.DiagnosticSource" version="4.4.0" targetFramework="net461" />
+</packages>
