+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/ADFSLogin.cs
