+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/DOCValidationTests/CreateUser.cs
