+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/DOCValidationTests/OpenMobile.cs
