+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/Dashboard.cs
@@ -7,7 +7,7 @@ using System.Security;
 namespace Microsoft.Dynamics365.UIAutomation.Sample
 {
     [TestClass]
-    public class CreateAccountTests
+    public class Dashboard
     {
 
         private readonly SecureString _username = System.Configuration.ConfigurationManager.AppSettings["OnlineUsername"].ToSecureString();
@@ -15,29 +15,16 @@ namespace Microsoft.Dynamics365.UIAutomation.Sample
         private readonly Uri _xrmUri = new Uri(System.Configuration.ConfigurationManager.AppSettings["OnlineCrmUrl"].ToString());
 
         [TestMethod]
-        public void TestCreateNewAccount()
+        public void TestSwitchDashboard()
         {
             using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
             {
                 xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
                 xrmBrowser.GuidedHelp.CloseGuidedHelp();
 
-                xrmBrowser.ThinkTime(500);
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Accounts");
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Pulpity nawigacyjne");
 
-                xrmBrowser.ThinkTime(2000);
-                xrmBrowser.Grid.SwitchView("Active Accounts");
-
-                xrmBrowser.ThinkTime(1000);
-                xrmBrowser.CommandBar.ClickCommand("New");
-
-                xrmBrowser.ThinkTime(4000);
-                xrmBrowser.Entity.SetValue("name", "Test API Account");
-                xrmBrowser.Entity.SetValue("telephone1", "555-555-5555");
-                xrmBrowser.Entity.SetValue("websiteurl", "https://easyrepro.crm.dynamics.com");
-
-                xrmBrowser.CommandBar.ClickCommand("Save & Close");
-                xrmBrowser.ThinkTime(2000);
+                xrmBrowser.Dashboard.SelectDashBoard("Pulpit nawigacyjny wydajności sprzedaży");
             }
         }
     }
