+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/GlobalSearch.cs
