+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/OpenAccount.cs
@@ -1,38 +1,36 @@
-﻿using System;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.Security;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Microsoft.Dynamics365.UIAutomation.Api;
 using Microsoft.Dynamics365.UIAutomation.Browser;
+using System;
+using System.Security;
 
 namespace Microsoft.Dynamics365.UIAutomation.Sample
 {
     [TestClass]
-    public class UpdateAccount
+    public class OpenAccount
     {
+
         private readonly SecureString _username = System.Configuration.ConfigurationManager.AppSettings["OnlineUsername"].ToSecureString();
         private readonly SecureString _password = System.Configuration.ConfigurationManager.AppSettings["OnlinePassword"].ToSecureString();
         private readonly Uri _xrmUri = new Uri(System.Configuration.ConfigurationManager.AppSettings["OnlineCrmUrl"].ToString());
 
         [TestMethod]
-        public void TestUpdateAccount()
+        public void TestOpenActiveAccount()
         {
             using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
             {
                 xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
+
                 xrmBrowser.GuidedHelp.CloseGuidedHelp();
+
                 xrmBrowser.ThinkTime(500);
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Accounts");
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Konta");
 
                 xrmBrowser.ThinkTime(2000);
-                xrmBrowser.Grid.SwitchView("Active Accounts");
+                xrmBrowser.Grid.SwitchView("Aktywne konta");
 
                 xrmBrowser.ThinkTime(1000);
                 xrmBrowser.Grid.OpenRecord(0);
-
-                xrmBrowser.ThinkTime(1000);
-                xrmBrowser.Entity.SelectLookup("parentaccountid", 0);
-                xrmBrowser.Entity.Save();
-                xrmBrowser.ThinkTime(2000);
             }
         }
     }
