+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/OpenCase.cs
@@ -1,39 +1,38 @@
 ﻿using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.Security;
 using Microsoft.Dynamics365.UIAutomation.Browser;
 using Microsoft.Dynamics365.UIAutomation.Api;
-using System.Security;
 
 namespace Microsoft.Dynamics365.UIAutomation.Sample
 {
     [TestClass]
-    public class UpdateOpportunity
+    public class OpenCase
     {
         private readonly SecureString _username = System.Configuration.ConfigurationManager.AppSettings["OnlineUsername"].ToSecureString();
         private readonly SecureString _password = System.Configuration.ConfigurationManager.AppSettings["OnlinePassword"].ToSecureString();
         private readonly Uri _xrmUri = new Uri(System.Configuration.ConfigurationManager.AppSettings["OnlineCrmUrl"].ToString());
 
         [TestMethod]
-        public void TestUpdateOpportunity()
+        public void TestOpenCase()
         {
             using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
             {
                 xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
+
                 xrmBrowser.GuidedHelp.CloseGuidedHelp();
 
                 xrmBrowser.ThinkTime(500);
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Opportunities");
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Konta");
 
-                xrmBrowser.ThinkTime(200);
-                xrmBrowser.Grid.SwitchView("Open Opportunities");
-
-
-                xrmBrowser.ThinkTime(1000);
+                xrmBrowser.ThinkTime(3000);
                 xrmBrowser.Grid.OpenRecord(0);
+                xrmBrowser.Navigation.OpenRelated("Sprawy");
 
-                xrmBrowser.Entity.SetValue("description", "Testing the update api for Opportunity");
+                xrmBrowser.Related.SwitchView("Aktywne sprawy");
 
-                xrmBrowser.Entity.Save();
+                xrmBrowser.ThinkTime(2000);
+                xrmBrowser.Related.OpenGridRow(0);
             }
         }
     }
