+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/OpenLead.cs
@@ -1,47 +1,32 @@
-﻿using System;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Microsoft.Dynamics365.UIAutomation.Api;
-using System.Security;
 using Microsoft.Dynamics365.UIAutomation.Browser;
+using System;
+using System.Security;
 
 namespace Microsoft.Dynamics365.UIAutomation.Sample
 {
     [TestClass]
-    public class CreateCase
+    public class OpenLead
     {
+
         private readonly SecureString _username = System.Configuration.ConfigurationManager.AppSettings["OnlineUsername"].ToSecureString();
         private readonly SecureString _password = System.Configuration.ConfigurationManager.AppSettings["OnlinePassword"].ToSecureString();
         private readonly Uri _xrmUri = new Uri(System.Configuration.ConfigurationManager.AppSettings["OnlineCrmUrl"].ToString());
 
         [TestMethod]
-        public void TestCreateNewCase()
+        public void TestOpenActiveLead()
         {
-
             using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
             {
                 xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
-
                 xrmBrowser.GuidedHelp.CloseGuidedHelp();
-
-                xrmBrowser.ThinkTime(500);
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Accounts");
-
-                xrmBrowser.ThinkTime(3000);
+                
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Potencjalni klienci");
+                
+                xrmBrowser.Grid.SwitchView("Wszyscy potencjalni klienci");
+                
                 xrmBrowser.Grid.OpenRecord(0);
-                xrmBrowser.Navigation.OpenRelated("Cases");
-
-                xrmBrowser.Related.SwitchView("Active Cases");
-                xrmBrowser.ThinkTime(2000);
-
-                xrmBrowser.Related.ClickCommand("ADD NEW CASE");
-                xrmBrowser.ThinkTime(2000);
-
-
-                xrmBrowser.QuickCreate.SetValue("title", "Test API Case");
-
-                xrmBrowser.QuickCreate.Save();
-                xrmBrowser.ThinkTime(10000);
-
             }
         }
     }
