+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/OpenOpportunity.cs
@@ -1,41 +1,34 @@
-﻿using System;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.Security;
-using Microsoft.Dynamics365.UIAutomation.Browser;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Microsoft.Dynamics365.UIAutomation.Api;
+using Microsoft.Dynamics365.UIAutomation.Browser;
+using System;
+using System.Security;
 
 namespace Microsoft.Dynamics365.UIAutomation.Sample
 {
     [TestClass]
-    public class OpenCase
+    public class OpenOpportunity
     {
+
         private readonly SecureString _username = System.Configuration.ConfigurationManager.AppSettings["OnlineUsername"].ToSecureString();
         private readonly SecureString _password = System.Configuration.ConfigurationManager.AppSettings["OnlinePassword"].ToSecureString();
         private readonly Uri _xrmUri = new Uri(System.Configuration.ConfigurationManager.AppSettings["OnlineCrmUrl"].ToString());
 
         [TestMethod]
-        public void TestOpenCase()
+        public void TestOpenActiveOpportunity()
         {
             using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
             {
                 xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
-
                 xrmBrowser.GuidedHelp.CloseGuidedHelp();
 
                 xrmBrowser.ThinkTime(500);
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Accounts");
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Szanse sprzedaży");
 
-                xrmBrowser.ThinkTime(3000);
+                xrmBrowser.ThinkTime(2000);
+                xrmBrowser.Grid.SwitchView("Otwarte szanse sprzedaży");
+
                 xrmBrowser.Grid.OpenRecord(0);
-                xrmBrowser.Navigation.OpenRelated("Cases");
-
-                xrmBrowser.Related.SwitchView("Active Cases");
-
-                xrmBrowser.ThinkTime(2000);
-                xrmBrowser.Related.OpenGridRow(0);
-                xrmBrowser.ThinkTime(2000);
-
-
             }
         }
     }
