+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/QuickCreate.cs
@@ -7,29 +7,26 @@ using System.Security;
 namespace Microsoft.Dynamics365.UIAutomation.Sample
 {
     [TestClass]
-    public class OpenOpportunity
+    public class QuickCreate
     {
-
         private readonly SecureString _username = System.Configuration.ConfigurationManager.AppSettings["OnlineUsername"].ToSecureString();
         private readonly SecureString _password = System.Configuration.ConfigurationManager.AppSettings["OnlinePassword"].ToSecureString();
         private readonly Uri _xrmUri = new Uri(System.Configuration.ConfigurationManager.AppSettings["OnlineCrmUrl"].ToString());
 
         [TestMethod]
-        public void TestOpenActiveOpportunity()
+        public void TestOpenCreate()
         {
             using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
             {
                 xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
                 xrmBrowser.GuidedHelp.CloseGuidedHelp();
 
-                xrmBrowser.ThinkTime(500);
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Opportunities");
+                xrmBrowser.Navigation.QuickCreate("Kontakt");
 
-                xrmBrowser.ThinkTime(2000);
-                xrmBrowser.Grid.SwitchView("Open Opportunities");
+                xrmBrowser.QuickCreate.SetValue("lastname", "Contact");
+                xrmBrowser.QuickCreate.SetValue("firstname", "Test Quick Create");
 
-                xrmBrowser.ThinkTime(1000);
-                xrmBrowser.Grid.OpenRecord(0);
+                xrmBrowser.QuickCreate.Save();
             }
         }
     }
