+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/Reports.cs
@@ -3,11 +3,12 @@ using Microsoft.Dynamics365.UIAutomation.Api;
 using Microsoft.Dynamics365.UIAutomation.Browser;
 using System;
 using System.Security;
+using OpenQA.Selenium;
 
 namespace Microsoft.Dynamics365.UIAutomation.Sample
 {
     [TestClass]
-    public class Views
+    public class Reports
     {
 
         private readonly SecureString _username = System.Configuration.ConfigurationManager.AppSettings["OnlineUsername"].ToSecureString();
@@ -15,67 +16,84 @@ namespace Microsoft.Dynamics365.UIAutomation.Sample
         private readonly Uri _xrmUri = new Uri(System.Configuration.ConfigurationManager.AppSettings["OnlineCrmUrl"].ToString());
 
         [TestMethod]
-        public void TestSortGridRow()
+        public void TestAccountOverviewReport()
         {
             using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
             {
                 xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
                 xrmBrowser.GuidedHelp.CloseGuidedHelp();
 
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Accounts");
-                xrmBrowser.Grid.SwitchView("Active Accounts");
-                xrmBrowser.Grid.Sort("Account Name");
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Raporty");
+
+                xrmBrowser.Grid.Search("Przegląd klientów");
+
+                xrmBrowser.Grid.OpenRecord(0);
+
+                xrmBrowser.Report.RunReport(null);
+            }
+        }
+
+        [TestMethod]
+        public void TestAccountOverviewFromGrid()
+        {
+            using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
+            {
+                xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
+                xrmBrowser.GuidedHelp.CloseGuidedHelp();
+
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Konta");
+
+                xrmBrowser.CommandBar.ClickCommand("Uruchom raport", "Przegląd klientów");
+
+                xrmBrowser.Dialogs.RunReport(XrmDialogPage.ReportRecords.AllRecords);
+            }
+        }
+        [TestMethod]
+        public void TestAccountOverviewFromRecord()
+        {
+            using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
+            {
+                xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
+                xrmBrowser.GuidedHelp.CloseGuidedHelp();
+
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Konta");
+
+                xrmBrowser.Grid.OpenRecord(0);
+
+                xrmBrowser.CommandBar.ClickCommand("Uruchom raport", "Przegląd klientów", true);
+            }
+        }
+
+        [TestMethod]
+        public void TestAccountOverviewWithCriteriaReport()
+        {
+            using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
+            {
+                xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
+                xrmBrowser.GuidedHelp.CloseGuidedHelp();
+
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Raporty");
+
+                xrmBrowser.Grid.Search("Przegląd klientów");
+
+                xrmBrowser.Grid.OpenRecord(0);
+
+                xrmBrowser.Report.RunReport(SetFilterCriteria);
+
                 xrmBrowser.ThinkTime(2000);
 
             }
         }
 
-        [TestMethod]
-        public void TestFilterGridByLetter()
+        private static void SetFilterCriteria(ReportEventArgs args)
         {
-            using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
-            {
-                xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
-                xrmBrowser.GuidedHelp.CloseGuidedHelp();
+            //Modify Report Filter to change the default "Modified On" of 30 days to 15 days
+            var criteria = args.Driver.FindElement(By.XPath("id(\"CRM_FilteredAccountEFGRP0FFLD0CCVALLBL\")"));
+            criteria.Click();
 
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Accounts");
-                xrmBrowser.Grid.SwitchView("Active Accounts");
-                xrmBrowser.Grid.FilterByLetter('A');
-                xrmBrowser.ThinkTime(2000);
+            var input = args.Driver.FindElement(By.XPath("id(\"CRM_FilteredAccountEFGRP0FFLD0CCVALCTL\")"));
+            input.SendKeys("15", true);
 
-            }
-        }
-
-        [TestMethod]
-        public void TestFilterGridByAll()
-        {
-            using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
-            {
-                xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
-                xrmBrowser.GuidedHelp.CloseGuidedHelp();
-
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Leads");
-                xrmBrowser.Grid.SwitchView("Open Leads");
-                xrmBrowser.Grid.FilterByAll();
-                xrmBrowser.ThinkTime(2000);
-
-            }
-        }
-
-        [TestMethod]
-        public void TestEnableFilter()
-        {
-            using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
-            {
-                xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
-                xrmBrowser.GuidedHelp.CloseGuidedHelp();
-
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Leads");
-                xrmBrowser.Grid.SwitchView("Open Leads");
-                xrmBrowser.Grid.EnableFilter();
-                xrmBrowser.ThinkTime(2000);
-
-            }
         }
     }
 }
