+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/Telemetry.cs
