+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/TestSettings.cs
