+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/UpdateAccount.cs
@@ -1,46 +1,38 @@
 ﻿using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Security;
-using Microsoft.Dynamics365.UIAutomation.Browser;
 using Microsoft.Dynamics365.UIAutomation.Api;
+using Microsoft.Dynamics365.UIAutomation.Browser;
 
 namespace Microsoft.Dynamics365.UIAutomation.Sample
 {
     [TestClass]
-    public class UpdateCase
+    public class UpdateAccount
     {
         private readonly SecureString _username = System.Configuration.ConfigurationManager.AppSettings["OnlineUsername"].ToSecureString();
         private readonly SecureString _password = System.Configuration.ConfigurationManager.AppSettings["OnlinePassword"].ToSecureString();
         private readonly Uri _xrmUri = new Uri(System.Configuration.ConfigurationManager.AppSettings["OnlineCrmUrl"].ToString());
 
         [TestMethod]
-        public void TestUpdateCase()
+        public void TestUpdateAccount()
         {
             using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
             {
                 xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
-
                 xrmBrowser.GuidedHelp.CloseGuidedHelp();
 
                 xrmBrowser.ThinkTime(500);
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Accounts");
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Konta");
 
-                xrmBrowser.ThinkTime(3000);
+                xrmBrowser.ThinkTime(2000);
+                xrmBrowser.Grid.SwitchView("Aktywne konta");
+
+                xrmBrowser.ThinkTime(1000);
                 xrmBrowser.Grid.OpenRecord(0);
-                xrmBrowser.Navigation.OpenRelated("Cases");
 
-                xrmBrowser.Related.SwitchView("Active Cases");
-
-                xrmBrowser.ThinkTime(2000);
-                xrmBrowser.Related.OpenGridRow(0);
-                xrmBrowser.ThinkTime(2000);
-
-                xrmBrowser.Entity.SetValue(new OptionSet { Name = "caseorigincode", Value = "Email" });
+                xrmBrowser.ThinkTime(1000);
+                xrmBrowser.Entity.SelectLookup("parentaccountid", 0);
                 xrmBrowser.Entity.Save();
-                xrmBrowser.ThinkTime(10000);
-
-
-
             }
         }
     }
