+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/UpdateCase.cs
@@ -7,40 +7,36 @@ using Microsoft.Dynamics365.UIAutomation.Api;
 namespace Microsoft.Dynamics365.UIAutomation.Sample
 {
     [TestClass]
-    public class UpdateContact
+    public class UpdateCase
     {
         private readonly SecureString _username = System.Configuration.ConfigurationManager.AppSettings["OnlineUsername"].ToSecureString();
         private readonly SecureString _password = System.Configuration.ConfigurationManager.AppSettings["OnlinePassword"].ToSecureString();
         private readonly Uri _xrmUri = new Uri(System.Configuration.ConfigurationManager.AppSettings["OnlineCrmUrl"].ToString());
 
         [TestMethod]
-        public void TestUpdateContact()
+        public void TestUpdateCase()
         {
             using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
             {
                 xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
+
                 xrmBrowser.GuidedHelp.CloseGuidedHelp();
 
-                var perf = xrmBrowser.PerformanceCenter;
-
-                if (!perf.IsEnabled)
-                    perf.IsEnabled = true;
-
                 xrmBrowser.ThinkTime(500);
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Contacts");
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Konta");
+
+                xrmBrowser.ThinkTime(3000);
+                xrmBrowser.Grid.OpenRecord(0);
+                xrmBrowser.Navigation.OpenRelated("Sprawy");
+
+                xrmBrowser.Related.SwitchView("Aktywne sprawy");
 
                 xrmBrowser.ThinkTime(2000);
-                xrmBrowser.Grid.SwitchView("Active Contacts");
+                xrmBrowser.Related.OpenGridRow(0);
+                xrmBrowser.ThinkTime(2000);
 
-                xrmBrowser.ThinkTime(1000);
-                xrmBrowser.Grid.OpenRecord(0);
-
-                xrmBrowser.Entity.SetValue("emailaddress1", "testUpdate@contoso.com");
-                xrmBrowser.Entity.SetValue("mobilephone", "123-222-4444");
-                xrmBrowser.Entity.SetValue("birthdate", DateTime.Parse("12/2/1984"));
-
+                xrmBrowser.Entity.SetValue(new OptionSet { Name = "caseorigincode", Value = "Email" });
                 xrmBrowser.Entity.Save();
-
             }
         }
     }
