+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/UpdateContact.cs
@@ -1,39 +1,45 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Microsoft.Dynamics365.UIAutomation.Api;
-using Microsoft.Dynamics365.UIAutomation.Browser;
-using System;
+﻿using System;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Security;
+using Microsoft.Dynamics365.UIAutomation.Browser;
+using Microsoft.Dynamics365.UIAutomation.Api;
 
 namespace Microsoft.Dynamics365.UIAutomation.Sample
 {
     [TestClass]
-    public class OpenAccount
+    public class UpdateContact
     {
-
         private readonly SecureString _username = System.Configuration.ConfigurationManager.AppSettings["OnlineUsername"].ToSecureString();
         private readonly SecureString _password = System.Configuration.ConfigurationManager.AppSettings["OnlinePassword"].ToSecureString();
         private readonly Uri _xrmUri = new Uri(System.Configuration.ConfigurationManager.AppSettings["OnlineCrmUrl"].ToString());
 
         [TestMethod]
-        public void TestOpenActiveAccount()
+        public void TestUpdateContact()
         {
             using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
             {
                 xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
-
                 xrmBrowser.GuidedHelp.CloseGuidedHelp();
 
+                var perf = xrmBrowser.PerformanceCenter;
+
+                if (!perf.IsEnabled)
+                    perf.IsEnabled = true;
+
                 xrmBrowser.ThinkTime(500);
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Accounts");
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Kontakty");
 
                 xrmBrowser.ThinkTime(2000);
-                xrmBrowser.Grid.SwitchView("Active Accounts");
+                xrmBrowser.Grid.SwitchView("Aktywne kontakty");
 
                 xrmBrowser.ThinkTime(1000);
                 xrmBrowser.Grid.OpenRecord(0);
 
-                xrmBrowser.ThinkTime(2000);
+                xrmBrowser.Entity.SetValue("emailaddress1", "testUpdate@contoso.com");
+                xrmBrowser.Entity.SetValue("mobilephone", "123-222-4444");
+                xrmBrowser.Entity.SetValue("birthdate", DateTime.Parse("12/2/1984"));
 
+                xrmBrowser.Entity.Save();
             }
         }
     }
