+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/UpdateLead.cs
@@ -1,35 +1,37 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Microsoft.Dynamics365.UIAutomation.Api;
+﻿using System;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Microsoft.Dynamics365.UIAutomation.Browser;
-using System;
+using Microsoft.Dynamics365.UIAutomation.Api;
 using System.Security;
 
 namespace Microsoft.Dynamics365.UIAutomation.Sample
 {
     [TestClass]
-    public class QuickCreate
+    public class UpdateLead
     {
         private readonly SecureString _username = System.Configuration.ConfigurationManager.AppSettings["OnlineUsername"].ToSecureString();
         private readonly SecureString _password = System.Configuration.ConfigurationManager.AppSettings["OnlinePassword"].ToSecureString();
         private readonly Uri _xrmUri = new Uri(System.Configuration.ConfigurationManager.AppSettings["OnlineCrmUrl"].ToString());
 
         [TestMethod]
-        public void TestOpenCreate()
+        public void TestUpdateLead()
         {
             using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
             {
                 xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
                 xrmBrowser.GuidedHelp.CloseGuidedHelp();
 
-                xrmBrowser.Navigation.QuickCreate("Contact");
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Potencjalni klienci");
 
-                xrmBrowser.QuickCreate.SetValue("lastname", "Contact");
-                xrmBrowser.QuickCreate.SetValue("firstname", "Test Quick Create");
+                xrmBrowser.Grid.SwitchView("Wszyscy potencjalni klienci");
 
-                xrmBrowser.QuickCreate.Save();
+                xrmBrowser.Grid.OpenRecord(0);
 
-                xrmBrowser.ThinkTime(4000);
+               
+                xrmBrowser.Entity.SetValue("subject", "Update test API Lead");
+                xrmBrowser.Entity.SetValue("description", "Test lead updation with API commands");
 
+                xrmBrowser.Entity.Save();
             }
         }
     }
