+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/UploadTelemetry.cs
