+++ Microsoft.Dynamics365.UIAutomation.Sample.Shared/Views.cs
@@ -3,12 +3,11 @@ using Microsoft.Dynamics365.UIAutomation.Api;
 using Microsoft.Dynamics365.UIAutomation.Browser;
 using System;
 using System.Security;
-using OpenQA.Selenium;
 
 namespace Microsoft.Dynamics365.UIAutomation.Sample
 {
     [TestClass]
-    public class Reports
+    public class Views
     {
 
         private readonly SecureString _username = System.Configuration.ConfigurationManager.AppSettings["OnlineUsername"].ToSecureString();
@@ -16,94 +15,63 @@ namespace Microsoft.Dynamics365.UIAutomation.Sample
         private readonly Uri _xrmUri = new Uri(System.Configuration.ConfigurationManager.AppSettings["OnlineCrmUrl"].ToString());
 
         [TestMethod]
-        public void TestAccountOverviewReport()
+        public void TestSortGridRow()
         {
             using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
             {
                 xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
                 xrmBrowser.GuidedHelp.CloseGuidedHelp();
 
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Reports");
-
-                xrmBrowser.Grid.Search("Account Overview");
-
-                xrmBrowser.Grid.OpenRecord(0);
-
-                xrmBrowser.Report.RunReport(null);
-
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Konta");
+                xrmBrowser.Grid.SwitchView("Aktywne konta");
+                xrmBrowser.Grid.Sort("Nazwa konta");
                 xrmBrowser.ThinkTime(2000);
-
             }
         }
 
         [TestMethod]
-        public void TestAccountOverviewFromGrid()
+        public void TestFilterGridByLetter()
         {
             using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
             {
                 xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
                 xrmBrowser.GuidedHelp.CloseGuidedHelp();
 
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Accounts");
-
-                xrmBrowser.CommandBar.ClickCommand("Run Report", "Account Overview");
-
-                xrmBrowser.Dialogs.RunReport(XrmDialogPage.ReportRecords.AllRecords);
-
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Konta");
+                xrmBrowser.Grid.SwitchView("Aktywne konta");
+                xrmBrowser.Grid.FilterByLetter('A');
                 xrmBrowser.ThinkTime(2000);
-
-            }
-        }
-        [TestMethod]
-        public void TestAccountOverviewFromRecord()
-        {
-            using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
-            {
-                xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
-                xrmBrowser.GuidedHelp.CloseGuidedHelp();
-
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Accounts");
-
-                xrmBrowser.Grid.OpenRecord(0);
-
-                xrmBrowser.CommandBar.ClickCommand("Run Report", "Account Overview", true);
-
-
-                xrmBrowser.ThinkTime(2000);
-
             }
         }
 
         [TestMethod]
-        public void TestAccountOverviewWithCriteriaReport()
+        public void TestFilterGridByAll()
         {
             using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
             {
                 xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
                 xrmBrowser.GuidedHelp.CloseGuidedHelp();
 
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Reports");
-
-                xrmBrowser.Grid.Search("Account Overview");
-
-                xrmBrowser.Grid.OpenRecord(0);
-
-                xrmBrowser.Report.RunReport(SetFilterCriteria);
-
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Potencjalni klienci");
+                xrmBrowser.Grid.SwitchView("Otwarci potencjalni klienci");
+                xrmBrowser.Grid.FilterByAll();
                 xrmBrowser.ThinkTime(2000);
-
             }
         }
 
-        public void SetFilterCriteria(ReportEventArgs args)
+        [TestMethod]
+        public void TestEnableFilter()
         {
-            //Modify Report Filter to change the default "Modified On" of 30 days to 15 days
-            var criteria = args.Driver.FindElement(By.XPath("id(\"CRM_FilteredAccountEFGRP0FFLD0CCVALLBL\")"));
-            criteria.Click();
+            using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
+            {
+                xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
+                xrmBrowser.GuidedHelp.CloseGuidedHelp();
 
-            var input = args.Driver.FindElement(By.XPath("id(\"CRM_FilteredAccountEFGRP0FFLD0CCVALCTL\")"));
-            input.SendKeys("15", true);
-
+                xrmBrowser.Navigation.OpenSubArea("Sprzedaż", "Potencjalni klienci");
+                xrmBrowser.Grid.SwitchView("Otwarci potencjalni klienci");
+                xrmBrowser.Grid.EnableFilter();
+                xrmBrowser.ThinkTime(2000);
+            }
         }
     }
 }
