--- Microsoft.Dynamics365.UIAutomation.Sample/CreateContact.cs
@@ -1,54 +1,0 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Microsoft.Dynamics365.UIAutomation.Api;
-using Microsoft.Dynamics365.UIAutomation.Browser;
-using System;
-using System.Collections.Generic;
-using System.Security;
-
-namespace Microsoft.Dynamics365.UIAutomation.Sample
-{
-    [TestClass]
-    public class CreateContact
-    {
-
-        private readonly SecureString _username = System.Configuration.ConfigurationManager.AppSettings["OnlineUsername"].ToSecureString();
-        private readonly SecureString _password = System.Configuration.ConfigurationManager.AppSettings["OnlinePassword"].ToSecureString();
-        private readonly Uri _xrmUri = new Uri(System.Configuration.ConfigurationManager.AppSettings["OnlineCrmUrl"].ToString());
-
-        [TestMethod]
-        public void TestCreateNewContact()
-        {
-            using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
-            {
-                xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
-                xrmBrowser.GuidedHelp.CloseGuidedHelp();
-
-                xrmBrowser.ThinkTime(500);
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Contacts");
-
-                xrmBrowser.ThinkTime(2000);
-                xrmBrowser.Grid.SwitchView("Active Contacts");
-
-                xrmBrowser.ThinkTime(1000);
-                xrmBrowser.CommandBar.ClickCommand("New");
-
-                xrmBrowser.ThinkTime(5000);
-
-                var fields = new List<Field>
-                {
-                    new Field() {Id = "firstname", Value = "Test"},
-                    new Field() {Id = "lastname", Value = "Contact"}
-                };
-                xrmBrowser.Entity.SetValue(new CompositeControl() {Id = "fullname", Fields = fields});
-                xrmBrowser.Entity.SetValue("emailaddress1", "test@contoso.com");
-                xrmBrowser.Entity.SetValue("mobilephone", "555-555-5555");
-                xrmBrowser.Entity.SetValue("birthdate", DateTime.Parse("11/1/1980"));
-                xrmBrowser.Entity.SetValue(new OptionSet {Name = "preferredcontactmethodcode", Value = "Email"});
-
-                xrmBrowser.CommandBar.ClickCommand("Save");
-                xrmBrowser.ThinkTime(5000);
-
-            }
-        }
-    }
-}
