--- Microsoft.Dynamics365.UIAutomation.Sample/CreateLead.cs
@@ -1,48 +1,0 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Microsoft.Dynamics365.UIAutomation.Api;
-using Microsoft.Dynamics365.UIAutomation.Browser;
-using System;
-using System.Collections.Generic;
-using System.Security;
-
-namespace Microsoft.Dynamics365.UIAutomation.Sample
-{
-    [TestClass]
-    public class CreateLead
-    {
-
-        private readonly SecureString _username = System.Configuration.ConfigurationManager.AppSettings["OnlineUsername"].ToSecureString();
-        private readonly SecureString _password = System.Configuration.ConfigurationManager.AppSettings["OnlinePassword"].ToSecureString();
-        private readonly Uri _xrmUri = new Uri(System.Configuration.ConfigurationManager.AppSettings["OnlineCrmUrl"].ToString());
-
-        [TestMethod]
-        public void TestCreateNewLead()
-        {
-            using (var xrmBrowser = new XrmBrowser(TestSettings.Options))
-            {
-                xrmBrowser.LoginPage.Login(_xrmUri, _username, _password);
-                xrmBrowser.GuidedHelp.CloseGuidedHelp();
-                
-                xrmBrowser.Navigation.OpenSubArea("Sales", "Leads");
-                
-                xrmBrowser.Grid.SwitchView("All Leads");
-                
-                xrmBrowser.CommandBar.ClickCommand("New");
-
-                xrmBrowser.ThinkTime(2000);
-                List<Field> fields = new List<Field>
-                {
-                    new Field() {Id = "firstname", Value = "Test"},
-                    new Field() {Id = "lastname", Value = "Lead"}
-                };
-                xrmBrowser.Entity.SetValue("subject", "Test API Lead");
-                xrmBrowser.Entity.SetValue(new CompositeControl() { Id = "fullname", Fields = fields });
-                xrmBrowser.Entity.SetValue("mobilephone", "555-555-5555");
-                xrmBrowser.Entity.SetValue("description", "Test lead creation with API commands");
-
-                xrmBrowser.CommandBar.ClickCommand("Save");
-                xrmBrowser.ThinkTime(2000);
-            }
-        }
-    }
-}
