+++ UIAutomation.sln
@@ -3,20 +3,25 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27130.2027
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Dynamics365.UIAutomation.Sample", "Microsoft.Dynamics365.UIAutomation.Sample\Microsoft.Dynamics365.UIAutomation.Sample.csproj", "{A075B87E-DCA5-4A2A-909B-932E2F5F98B4}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Dynamics365.UIAutomation.Api", "Microsoft.Dynamics365.UIAutomation.Api\Microsoft.Dynamics365.UIAutomation.Api.csproj", "{DFE2C141-F1AA-4D4B-8DF8-0FB93023CF45}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Dynamics365.UIAutomation.Browser", "Microsoft.Dynamics365.UIAutomation.Browser\Microsoft.Dynamics365.UIAutomation.Browser.csproj", "{C9748803-F3CB-4531-B313-BC24D4D0BF0C}"
 EndProject
 Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Microsoft.Dynamics365.UIAutomation.Api.Shared", "Microsoft.Dynamics365.UIAutomation.Api.Shared\Microsoft.Dynamics365.UIAutomation.Api.Shared.shproj", "{632515D8-12C2-4A36-B900-B29D283933A6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Dynamics365.UIAutomation.Api.2016", "Microsoft.Dynamics365.UIAutomation.Api.2016\Microsoft.Dynamics365.UIAutomation.Api.2016.csproj", "{80A65998-62A8-4E7B-87A0-8492C16703D3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Dynamics365.UIAutomation.Api.OnPremise", "Microsoft.Dynamics365.UIAutomation.Api.2016\Microsoft.Dynamics365.UIAutomation.Api.OnPremise.csproj", "{80A65998-62A8-4E7B-87A0-8492C16703D3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Dynamics365.UIAutomation.Api.Online", "Microsoft.Dynamics365.UIAutomation.Api.365\Microsoft.Dynamics365.UIAutomation.Api.Online.csproj", "{73DCF3F6-DD2F-43B1-B795-4428EC0456D4}"
+EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Microsoft.Dynamics365.UIAutomation.Sample.Shared", "Microsoft.Dynamics365.UIAutomation.Sample.Shared\Microsoft.Dynamics365.UIAutomation.Sample.Shared.shproj", "{430C39E8-41DF-4E82-95C9-39153107A30F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Dynamics365.UIAutomation.Sample.OnPremise", "Microsoft.Dynamics365.UIAutomation.Sample.2016\Microsoft.Dynamics365.UIAutomation.Sample.OnPremise.csproj", "{B2829BBC-2DFC-4BEB-BCFA-7E48DC488C8E}"
 EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		Microsoft.Dynamics365.UIAutomation.Sample.Shared\Microsoft.Dynamics365.UIAutomation.Sample.Shared.projitems*{430c39e8-41df-4e82-95c9-39153107a30f}*SharedItemsImports = 13
 		Microsoft.Dynamics365.UIAutomation.Api.Shared\Microsoft.Dynamics365.UIAutomation.Api.Shared.projitems*{632515d8-12c2-4a36-b900-b29d283933a6}*SharedItemsImports = 13
+		Microsoft.Dynamics365.UIAutomation.Api.Shared\Microsoft.Dynamics365.UIAutomation.Api.Shared.projitems*{73dcf3f6-dd2f-43b1-b795-4428ec0456d4}*SharedItemsImports = 4
 		Microsoft.Dynamics365.UIAutomation.Api.Shared\Microsoft.Dynamics365.UIAutomation.Api.Shared.projitems*{80a65998-62a8-4e7b-87a0-8492c16703d3}*SharedItemsImports = 4
+		Microsoft.Dynamics365.UIAutomation.Sample.Shared\Microsoft.Dynamics365.UIAutomation.Sample.Shared.projitems*{b2829bbc-2dfc-4beb-bcfa-7e48dc488c8e}*SharedItemsImports = 4
 	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -24,18 +29,6 @@ Global
 		UnitTest|Any CPU = UnitTest|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A075B87E-DCA5-4A2A-909B-932E2F5F98B4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A075B87E-DCA5-4A2A-909B-932E2F5F98B4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A075B87E-DCA5-4A2A-909B-932E2F5F98B4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A075B87E-DCA5-4A2A-909B-932E2F5F98B4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A075B87E-DCA5-4A2A-909B-932E2F5F98B4}.UnitTest|Any CPU.ActiveCfg = UnitTest|Any CPU
-		{A075B87E-DCA5-4A2A-909B-932E2F5F98B4}.UnitTest|Any CPU.Build.0 = UnitTest|Any CPU
-		{DFE2C141-F1AA-4D4B-8DF8-0FB93023CF45}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DFE2C141-F1AA-4D4B-8DF8-0FB93023CF45}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DFE2C141-F1AA-4D4B-8DF8-0FB93023CF45}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DFE2C141-F1AA-4D4B-8DF8-0FB93023CF45}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DFE2C141-F1AA-4D4B-8DF8-0FB93023CF45}.UnitTest|Any CPU.ActiveCfg = Release|Any CPU
-		{DFE2C141-F1AA-4D4B-8DF8-0FB93023CF45}.UnitTest|Any CPU.Build.0 = Release|Any CPU
 		{C9748803-F3CB-4531-B313-BC24D4D0BF0C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C9748803-F3CB-4531-B313-BC24D4D0BF0C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C9748803-F3CB-4531-B313-BC24D4D0BF0C}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -48,6 +41,18 @@ Global
 		{80A65998-62A8-4E7B-87A0-8492C16703D3}.Release|Any CPU.Build.0 = Release|Any CPU
 		{80A65998-62A8-4E7B-87A0-8492C16703D3}.UnitTest|Any CPU.ActiveCfg = Release|Any CPU
 		{80A65998-62A8-4E7B-87A0-8492C16703D3}.UnitTest|Any CPU.Build.0 = Release|Any CPU
+		{73DCF3F6-DD2F-43B1-B795-4428EC0456D4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{73DCF3F6-DD2F-43B1-B795-4428EC0456D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{73DCF3F6-DD2F-43B1-B795-4428EC0456D4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{73DCF3F6-DD2F-43B1-B795-4428EC0456D4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{73DCF3F6-DD2F-43B1-B795-4428EC0456D4}.UnitTest|Any CPU.ActiveCfg = Release|Any CPU
+		{73DCF3F6-DD2F-43B1-B795-4428EC0456D4}.UnitTest|Any CPU.Build.0 = Release|Any CPU
+		{B2829BBC-2DFC-4BEB-BCFA-7E48DC488C8E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B2829BBC-2DFC-4BEB-BCFA-7E48DC488C8E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B2829BBC-2DFC-4BEB-BCFA-7E48DC488C8E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B2829BBC-2DFC-4BEB-BCFA-7E48DC488C8E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B2829BBC-2DFC-4BEB-BCFA-7E48DC488C8E}.UnitTest|Any CPU.ActiveCfg = Release|Any CPU
+		{B2829BBC-2DFC-4BEB-BCFA-7E48DC488C8E}.UnitTest|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
